Moves tx allowlist check to allow replay protection optimizations
