+++ crates/namada/src/vm/wasm/run.rs
@@ -9,8 +9,8 @@ use borsh::BorshDeserialize;
 use namada_core::validity_predicate::VpSentinel;
 use namada_gas::{GasMetering, TxGasMeter, WASM_MEMORY_PAGE_GAS};
 use namada_state::write_log::StorageModification;
-use namada_state::{DBIter, State, StateRead, StorageHasher, DB};
-use namada_tx::data::TxSentinel;
+use namada_state::{DBIter, State, StateRead, StorageHasher, StorageRead, DB};
+use namada_tx::data::{TxSentinel, TxType};
 use namada_tx::{Commitment, Section, Tx};
 use parity_wasm::elements;
 use thiserror::Error;
@@ -84,10 +84,34 @@ pub enum Error {
     ConversionError(String),
     #[error("Invalid transaction signature")]
     InvalidTxSignature,
+    #[error("Tx is not allowed in allowlist parameter")]
+    DisallowedTx,
 }
 
 /// Result for functions that may fail
 pub type Result<T> = std::result::Result<T, Error>;
+
+// Returns [`Error::DisallowedTx`] when the given tx is inner (decrypted) tx
+// and its code `Hash` is not included in the `tx_allowlist` parameter.
+fn check_tx_allowed<S>(tx: &Tx, storage: &S) -> Result<()>
+where
+    S: StorageRead,
+{
+    if let TxType::Wrapper(_) = tx.header().tx_type {
+        if let Some(code_sec) = tx
+            .get_section(tx.code_sechash())
+            .and_then(|x| Section::code_sec(&x))
+        {
+            if crate::parameters::is_tx_allowed(storage, &code_sec.code.hash())
+                .map_err(|e| Error::LoadWasmCode(e.to_string()))?
+            {
+                return Ok(());
+            }
+        }
+        return Err(Error::DisallowedTx);
+    }
+    Ok(())
+}
 
 /// Execute a transaction code. Returns the set verifiers addresses requested by
 /// the transaction.
@@ -101,13 +125,18 @@ pub fn tx<S, CA>(
     tx_wasm_cache: &mut TxCache<CA>,
 ) -> Result<BTreeSet<Address>>
 where
-    S: StateRead + State,
+    S: StateRead + State + StorageRead,
     CA: 'static + WasmCacheAccess,
 {
     let tx_code = tx
         .get_section(tx.code_sechash())
         .and_then(|x| Section::code_sec(x.as_ref()))
         .ok_or(Error::MissingSection(tx.code_sechash().to_string()))?;
+
+    // Check if the tx code is allowed (to be done after the check on the code
+    // section commitment to let the replay protection mechanism run some
+    // optimizations)
+    check_tx_allowed(tx, state)?;
 
     // If the transaction code has a tag, ensure that the tag hash equals the
     // transaction code's hash.
@@ -1281,6 +1310,52 @@ mod tests {
         )
         .unwrap();
         assert!(!passed);
+    }
+
+    #[test]
+    fn test_apply_wasm_tx_allowlist() {
+        let mut state = TestState::default();
+
+        let tx_read_key = TestWasms::TxReadStorageKey.read_bytes();
+        // store the wasm code
+        let read_code_hash = Hash::sha256(&tx_read_key);
+        let code_len = (tx_read_key.len() as u64).serialize_to_vec();
+        let key = Key::wasm_code(&read_code_hash);
+        let len_key = Key::wasm_code_len(&read_code_hash);
+        state.write_bytes(&key, tx_read_key).unwrap();
+        state.write_bytes(&len_key, code_len).unwrap();
+
+        let mut tx = Tx::new(state.in_mem().chain_id.clone(), None);
+        tx.add_code_from_hash(read_code_hash, None)
+            .add_serialized_data(vec![]);
+
+        // Check that using a disallowed tx leads to an error
+        {
+            let allowlist = vec![format!("{}-bad", read_code_hash)];
+            crate::parameters::update_tx_allowlist_parameter(
+                &mut state, allowlist,
+            )
+            .unwrap();
+            state.commit_tx();
+
+            let result = check_tx_allowed(&tx, &state);
+            assert_matches!(result.unwrap_err(), Error::DisallowedTx);
+        }
+
+        // Check that using an allowed tx doesn't lead to `Error::DisallowedTx`
+        {
+            let allowlist = vec![read_code_hash.to_string()];
+            crate::parameters::update_tx_allowlist_parameter(
+                &mut state, allowlist,
+            )
+            .unwrap();
+            state.commit_tx();
+
+            let result = check_tx_allowed(&tx, &state);
+            if let Err(result) = result {
+                assert!(!matches!(result, Error::DisallowedTx));
+            }
+        }
     }
 
     fn execute_tx_with_code(tx_code: Vec<u8>) -> Result<BTreeSet<Address>> {
